[Refactoring] Move ExtractRepeatedExpr and ConvertToSwitchStmt to its own file
