+++ lib/Refactoring/CMakeLists.txt
@@ -1,9 +1,13 @@
 add_swift_host_library(swiftRefactoring STATIC
   CollapseNestedIfStmt.cpp
   ConvertStringConcatenationToInterpolation.cpp
+  ConvertToSwitchStmt.cpp
+  ExtractRepeatedExpr.cpp
   MoveMembersToExtension.cpp
   Refactoring.cpp
+  ExtractExprBase.cpp
   ReplaceBodiesWithFatalError.cpp
+  Utils.cpp
 )
 
 target_link_libraries(swiftRefactoring PRIVATE
