+++ lib/Refactoring/Refactoring.cpp
@@ -11,7 +11,10 @@
 //===----------------------------------------------------------------------===//
 
 #include "swift/Refactoring/Refactoring.h"
+#include "ContextFinder.h"
+#include "ExtractExprBase.h"
 #include "RefactoringActions.h"
+#include "Utils.h"
 #include "swift/AST/ASTContext.h"
 #include "swift/AST/ASTPrinter.h"
 #include "swift/AST/Decl.h"
@@ -43,52 +46,6 @@ using namespace swift;
 using namespace swift::ide;
 using namespace swift::index;
 using namespace swift::refactoring;
-
-class ContextFinder : public SourceEntityWalker {
-  SourceFile &SF;
-  ASTContext &Ctx;
-  SourceManager &SM;
-  SourceRange Target;
-  std::function<bool(ASTNode)> IsContext;
-  SmallVector<ASTNode, 4> AllContexts;
-  bool contains(ASTNode Enclosing) {
-    auto Result = SM.rangeContainsRespectingReplacedRanges(
-        Enclosing.getSourceRange(), Target);
-    if (Result && IsContext(Enclosing)) {
-      AllContexts.push_back(Enclosing);
-    }
-    return Result;
-  }
-
-public:
-  ContextFinder(SourceFile &SF, ASTNode TargetNode,
-                std::function<bool(ASTNode)> IsContext =
-                  [](ASTNode N) { return true; }) :
-                  SF(SF), Ctx(SF.getASTContext()), SM(Ctx.SourceMgr),
-                  Target(TargetNode.getSourceRange()), IsContext(IsContext) {}
-
-  ContextFinder(SourceFile &SF, SourceLoc TargetLoc,
-                std::function<bool(ASTNode)> IsContext =
-                  [](ASTNode N) { return true; }) :
-                  SF(SF), Ctx(SF.getASTContext()), SM(Ctx.SourceMgr),
-                  Target(TargetLoc), IsContext(IsContext) {
-                    assert(TargetLoc.isValid() && "Invalid loc to find");
-  }
-
-  // Only need expansions for the expands refactoring, but we
-  // skip nodes that don't contain the passed location anyway.
-  virtual MacroWalking getMacroWalkingBehavior() const override {
-    return MacroWalking::ArgumentsAndExpansion;
-  }
-
-  bool walkToDeclPre(Decl *D, CharSourceRange Range) override { return contains(D); }
-  bool walkToStmtPre(Stmt *S) override { return contains(S); }
-  bool walkToExprPre(Expr *E) override { return contains(E); }
-  void resolve() { walk(SF); }
-  ArrayRef<ASTNode> getContexts() const {
-    return llvm::makeArrayRef(AllContexts);
-  }
-};
 
 class Renamer {
 protected:
@@ -981,39 +938,12 @@ StringRef getDefaultPreferredName(RefactoringKind Kind) {
   }
 }
 
-enum class CannotExtractReason {
-  Literal,
-  VoidType,
-};
-
-class ExtractCheckResult {
-  bool KnownFailure;
-  SmallVector<CannotExtractReason, 2> AllReasons;
-
-public:
-  ExtractCheckResult(): KnownFailure(true) {}
-  ExtractCheckResult(ArrayRef<CannotExtractReason> AllReasons):
-    KnownFailure(false), AllReasons(AllReasons.begin(), AllReasons.end()) {}
-  bool success() { return success({}); }
-  bool success(ArrayRef<CannotExtractReason> ExpectedReasons) {
-    if (KnownFailure)
-      return false;
-    bool Result = true;
-
-    // Check if any reasons aren't covered by the list of expected reasons
-    // provided by the client.
-    for (auto R: AllReasons) {
-      Result &= llvm::is_contained(ExpectedReasons, R);
-    }
-    return Result;
-  }
-};
-
 /// Check whether a given range can be extracted.
 /// Return true on successful condition checking,.
 /// Return false on failed conditions.
-ExtractCheckResult checkExtractConditions(const ResolvedRangeInfo &RangeInfo,
-                                          DiagnosticEngine &DiagEngine) {
+ExtractCheckResult
+swift::refactoring::checkExtractConditions(const ResolvedRangeInfo &RangeInfo,
+                                           DiagnosticEngine &DiagEngine) {
   SmallVector<CannotExtractReason, 2> AllReasons;
   // If any declared declaration is referred out of the given range, return false.
   auto Declared = RangeInfo.DeclaredDecls;
@@ -1129,36 +1059,6 @@ isApplicable(const ResolvedRangeInfo &Info, DiagnosticEngine &Diag) {
   }
   }
   llvm_unreachable("unhandled kind");
-}
-
-static StringRef correctNameInternal(ASTContext &Ctx, StringRef Name,
-                                     ArrayRef<ValueDecl*> AllVisibles) {
-  // If we find the collision.
-  bool FoundCollision = false;
-
-  // The suffixes we cannot use by appending to the original given name.
-  llvm::StringSet<> UsedSuffixes;
-  for (auto VD : AllVisibles) {
-    StringRef S = VD->getBaseName().userFacingName();
-    if (!S.startswith(Name))
-      continue;
-    StringRef Suffix = S.substr(Name.size());
-    if (Suffix.empty())
-      FoundCollision = true;
-    else
-      UsedSuffixes.insert(Suffix);
-  }
-  if (!FoundCollision)
-    return Name;
-
-  // Find the first suffix we can use.
-  std::string SuffixToUse;
-  for (unsigned I = 1; ; I ++) {
-    SuffixToUse = std::to_string(I);
-    if (UsedSuffixes.count(SuffixToUse) == 0)
-      break;
-  }
-  return Ctx.getIdentifier((llvm::Twine(Name) + SuffixToUse).str()).str();
 }
 
 static StringRef correctNewDeclName(DeclContext *DC, StringRef Name) {
@@ -1440,216 +1340,6 @@ bool RefactoringActionExtractFunction::performChange() {
   return false;
 }
 
-class RefactoringActionExtractExprBase {
-  SourceFile *TheFile;
-  ResolvedRangeInfo RangeInfo;
-  DiagnosticEngine &DiagEngine;
-  const bool ExtractRepeated;
-  StringRef PreferredName;
-  SourceEditConsumer &EditConsumer;
-
-  ASTContext &Ctx;
-  SourceManager &SM;
-
-public:
-  RefactoringActionExtractExprBase(SourceFile *TheFile,
-                                   ResolvedRangeInfo RangeInfo,
-                                   DiagnosticEngine &DiagEngine,
-                                   bool ExtractRepeated,
-                                   StringRef PreferredName,
-                                   SourceEditConsumer &EditConsumer) :
-    TheFile(TheFile), RangeInfo(RangeInfo), DiagEngine(DiagEngine),
-    ExtractRepeated(ExtractRepeated), PreferredName(PreferredName),
-    EditConsumer(EditConsumer), Ctx(TheFile->getASTContext()),
-    SM(Ctx.SourceMgr){}
-  bool performChange();
-};
-
-/// This is to ensure all decl references in two expressions are identical.
-struct ReferenceCollector: public SourceEntityWalker {
-  SmallVector<ValueDecl*, 4> References;
-
-  ReferenceCollector(Expr *E) { walk(E); }
-  bool visitDeclReference(ValueDecl *D, CharSourceRange Range,
-                          TypeDecl *CtorTyRef, ExtensionDecl *ExtTyRef,
-                          Type T, ReferenceMetaData Data) override {
-    References.emplace_back(D);
-    return true;
-  }
-  bool operator==(const ReferenceCollector &Other) const {
-    if (References.size() != Other.References.size())
-      return false;
-    return std::equal(References.begin(), References.end(),
-                      Other.References.begin());
-  }
-};
-
-struct SimilarExprCollector: public SourceEntityWalker {
-  SourceManager &SM;
-
-  /// The expression under selection.
-  Expr *SelectedExpr;
-  ArrayRef<Token> AllTokens;
-  llvm::SetVector<Expr*> &Bucket;
-
-  /// The tokens included in the expression under selection.
-  ArrayRef<Token> SelectedTokens;
-
-  /// The referenced decls in the expression under selection.
-  ReferenceCollector SelectedReferences;
-
-  bool compareTokenContent(ArrayRef<Token> Left, ArrayRef<Token> Right) {
-    if (Left.size() != Right.size())
-      return false;
-    return std::equal(Left.begin(), Left.end(), Right.begin(),
-                      [](const Token &L, const Token& R) {
-                        return L.getText() == R.getText();
-                      });
-  }
-
-  /// Find all tokens included by an expression.
-  ArrayRef<Token> getExprSlice(Expr *E) {
-    return slice_token_array(AllTokens, E->getStartLoc(), E->getEndLoc());
-  }
-
-  SimilarExprCollector(SourceManager &SM, Expr *SelectedExpr,
-                       ArrayRef<Token> AllTokens,
-    llvm::SetVector<Expr*> &Bucket): SM(SM), SelectedExpr(SelectedExpr),
-    AllTokens(AllTokens), Bucket(Bucket),
-    SelectedTokens(getExprSlice(SelectedExpr)),
-    SelectedReferences(SelectedExpr){}
-
-  bool walkToExprPre(Expr *E) override {
-    // We don't extract implicit expressions.
-    if (E->isImplicit())
-      return true;
-    if (E->getKind() != SelectedExpr->getKind())
-      return true;
-
-    // First check the underlying token arrays have the same content.
-    if (compareTokenContent(getExprSlice(E), SelectedTokens)) {
-      ReferenceCollector CurrentReferences(E);
-
-      // Next, check the referenced decls are same.
-      if (CurrentReferences == SelectedReferences)
-        Bucket.insert(E);
-    }
-    return true;
-  }
-};
-
-bool RefactoringActionExtractExprBase::performChange() {
-  // Check if the new name is ok.
-  if (!Lexer::isIdentifier(PreferredName)) {
-    DiagEngine.diagnose(SourceLoc(), diag::invalid_name, PreferredName);
-    return true;
-  }
-
-  // Find the enclosing brace statement;
-  ContextFinder Finder(*TheFile, RangeInfo.ContainedNodes.front(),
-                       [](ASTNode N) { return N.isStmt(StmtKind::Brace); });
-
-  auto *SelectedExpr = RangeInfo.ContainedNodes[0].get<Expr*>();
-  Finder.resolve();
-  SourceLoc InsertLoc;
-  llvm::SetVector<ValueDecl*> AllVisibleDecls;
-  struct DeclCollector: public SourceEntityWalker {
-    llvm::SetVector<ValueDecl*> &Bucket;
-    DeclCollector(llvm::SetVector<ValueDecl*> &Bucket): Bucket(Bucket) {}
-    bool walkToDeclPre(Decl *D, CharSourceRange Range) override {
-      if (auto *VD = dyn_cast<ValueDecl>(D))
-        Bucket.insert(VD);
-      return true;
-    }
-  } Collector(AllVisibleDecls);
-
-  llvm::SetVector<Expr*> AllExpressions;
-
-  if (!Finder.getContexts().empty()) {
-
-    // Get the innermost brace statement.
-    auto BS = static_cast<BraceStmt*>(Finder.getContexts().back().get<Stmt*>());
-
-    // Collect all value decls inside the brace statement.
-    Collector.walk(BS);
-
-    if (ExtractRepeated) {
-      // Collect all expressions we are going to extract.
-      SimilarExprCollector(SM, SelectedExpr,
-                           slice_token_array(TheFile->getAllTokens(),
-                                             BS->getStartLoc(),
-                                             BS->getEndLoc()),
-                           AllExpressions).walk(BS);
-    } else {
-      AllExpressions.insert(SelectedExpr);
-    }
-
-    assert(!AllExpressions.empty() && "at least one expression is extracted.");
-    for (auto Ele : BS->getElements()) {
-      // Find the element that encloses the first expression under extraction.
-      if (SM.rangeContains(Ele.getSourceRange(),
-                           (*AllExpressions.begin())->getSourceRange())) {
-
-        // Insert before the enclosing element.
-        InsertLoc = Ele.getStartLoc();
-      }
-    }
-  }
-
-  // Complain about no inserting position.
-  if (InsertLoc.isInvalid()) {
-    DiagEngine.diagnose(SourceLoc(), diag::no_insert_position);
-    return true;
-  }
-
-  // Correct name if collision happens.
-  PreferredName = correctNameInternal(TheFile->getASTContext(), PreferredName,
-                                      AllVisibleDecls.getArrayRef());
-
-  // Print the type name of this expression.
-  SmallString<16> TyBuffer;
-
-  // We are not sure about the type of repeated expressions.
-  if (!ExtractRepeated) {
-    if (auto Ty = RangeInfo.getType()) {
-      llvm::raw_svector_ostream OS(TyBuffer);
-      OS << ": ";
-      Ty->getRValueType()->reconstituteSugar(true)->print(OS);
-    }
-  }
-
-  SmallString<64> DeclBuffer;
-  llvm::raw_svector_ostream OS(DeclBuffer);
-  unsigned StartOffset, EndOffset;
-  OS << tok::kw_let << " ";
-  StartOffset = DeclBuffer.size();
-  OS << PreferredName;
-  EndOffset = DeclBuffer.size();
-  OS << TyBuffer.str() <<  " = " << RangeInfo.ContentRange.str() << "\n";
-
-  NoteRegion DeclNameRegion{RefactoringRangeKind::BaseName,
-                            /*StartLine=*/1,
-                            /*StartColumn=*/StartOffset + 1,
-                            /*EndLine=*/1,
-                            /*EndColumn=*/EndOffset + 1,
-                            /*ArgIndex*/ llvm::None};
-
-  // Perform code change.
-  EditConsumer.accept(SM, InsertLoc, DeclBuffer.str(), {DeclNameRegion});
-
-  // Replace all occurrences of the extracted expression.
-  for (auto *E : AllExpressions) {
-    EditConsumer.accept(
-        SM, Lexer::getCharSourceRangeFromSourceRange(SM, E->getSourceRange()),
-        PreferredName,
-        {{RefactoringRangeKind::BaseName,
-          /*StartLine=*/1, /*StartColumn-*/ 1, /*EndLine=*/1,
-          /*EndColumn=*/static_cast<unsigned int>(PreferredName.size() + 1),
-          /*ArgIndex*/ llvm::None}});
-  }
-  return false;
-}
-
 bool RefactoringActionExtractExpr::
 isApplicable(const ResolvedRangeInfo &Info, DiagnosticEngine &Diag) {
   switch (Info.Kind) {
@@ -1673,28 +1363,6 @@ isApplicable(const ResolvedRangeInfo &Info, DiagnosticEngine &Diag) {
 bool RefactoringActionExtractExpr::performChange() {
   return RefactoringActionExtractExprBase(TheFile, RangeInfo,
                                           DiagEngine, false, PreferredName,
-                                          EditConsumer).performChange();
-}
-
-bool RefactoringActionExtractRepeatedExpr::
-isApplicable(const ResolvedRangeInfo &Info, DiagnosticEngine &Diag) {
-  switch (Info.Kind) {
-    case RangeKind::SingleExpression:
-      return checkExtractConditions(Info, Diag).
-        success({CannotExtractReason::Literal});
-    case RangeKind::PartOfExpression:
-    case RangeKind::SingleDecl:
-    case RangeKind::MultiTypeMemberDecl:
-    case RangeKind::SingleStatement:
-    case RangeKind::MultiStatement:
-    case RangeKind::Invalid:
-      return false;
-  }
-  llvm_unreachable("unhandled kind");
-}
-bool RefactoringActionExtractRepeatedExpr::performChange() {
-  return RefactoringActionExtractExprBase(TheFile, RangeInfo,
-                                          DiagEngine, true, PreferredName,
                                           EditConsumer).performChange();
 }
 
@@ -2072,293 +1740,6 @@ bool RefactoringActionConvertGuardExprToIfLetExpr::performChange() {
   auto ReplaceRange = RangeInfo.ContentRange;
   EditConsumer.accept(SM, ReplaceRange, DeclBuffer.str());
   
-  return false;
-}
-
-bool RefactoringActionConvertToSwitchStmt::
-isApplicable(const ResolvedRangeInfo &Info, DiagnosticEngine &Diag) {
-
-  class ConditionalChecker : public ASTWalker {
-  public:
-    bool ParamsUseSameVars = true;
-    bool ConditionUseOnlyAllowedFunctions = false;
-    StringRef ExpectName;
-
-    MacroWalking getMacroWalkingBehavior() const override {
-      return MacroWalking::Arguments;
-    }
-
-    PostWalkResult<Expr *> walkToExprPost(Expr *E) override {
-      if (E->getKind() != ExprKind::DeclRef)
-        return Action::Continue(E);
-      auto D = dyn_cast<DeclRefExpr>(E)->getDecl();
-      if (D->getKind() == DeclKind::Var || D->getKind() == DeclKind::Param)
-        ParamsUseSameVars = checkName(dyn_cast<VarDecl>(D));
-      if (D->getKind() == DeclKind::Func)
-        ConditionUseOnlyAllowedFunctions = checkName(dyn_cast<FuncDecl>(D));
-      if (allCheckPassed())
-        return Action::Continue(E);
-
-      return Action::Stop();
-    }
-
-    bool allCheckPassed() {
-      return ParamsUseSameVars && ConditionUseOnlyAllowedFunctions;
-    }
-
-  private:
-    bool checkName(VarDecl *VD) {
-      auto Name = VD->getName().str();
-      if (ExpectName.empty())
-        ExpectName = Name;
-      return Name == ExpectName;
-    }
-
-    bool checkName(FuncDecl *FD) {
-      const auto Name = FD->getBaseIdentifier().str();
-      return Name == "~="
-      || Name == "=="
-      || Name == "__derived_enum_equals"
-      || Name == "__derived_struct_equals"
-      || Name == "||"
-      || Name == "...";
-    }
-  };
-
-  class SwitchConvertable {
-  public:
-    SwitchConvertable(const ResolvedRangeInfo &Info) : Info(Info) { }
-
-    bool isApplicable() {
-      if (Info.Kind != RangeKind::SingleStatement)
-        return false;
-      if (!findIfStmt())
-        return false;
-      return checkEachCondition();
-    }
-
-  private:
-    const ResolvedRangeInfo &Info;
-    IfStmt *If = nullptr;
-    ConditionalChecker checker;
-
-    bool findIfStmt() {
-      if (Info.ContainedNodes.size() != 1)
-        return false;
-      if (auto S = Info.ContainedNodes.front().dyn_cast<Stmt*>())
-        If = dyn_cast<IfStmt>(S);
-      return If != nullptr;
-    }
-
-    bool checkEachCondition() {
-      checker = ConditionalChecker();
-      do {
-        if (!checkEachElement())
-          return false;
-      } while ((If = dyn_cast_or_null<IfStmt>(If->getElseStmt())));
-      return true;
-    }
-
-    bool checkEachElement() {
-      bool result = true;
-      auto ConditionalList = If->getCond();
-      for (auto Element : ConditionalList) {
-        result &= check(Element);
-      }
-      return result;
-    }
-
-    bool check(StmtConditionElement ConditionElement) {
-      if (ConditionElement.getKind() == StmtConditionElement::CK_Availability)
-        return false;
-      if (ConditionElement.getKind() == StmtConditionElement::CK_PatternBinding)
-        checker.ConditionUseOnlyAllowedFunctions = true;
-      ConditionElement.walk(checker);
-      return checker.allCheckPassed();
-    }
-  };
-  return SwitchConvertable(Info).isApplicable();
-}
-
-bool RefactoringActionConvertToSwitchStmt::performChange() {
-
-  class VarNameFinder : public ASTWalker {
-  public:
-    std::string VarName;
-
-    MacroWalking getMacroWalkingBehavior() const override {
-      return MacroWalking::Arguments;
-    }
-
-    PostWalkResult<Expr *> walkToExprPost(Expr *E) override {
-      if (E->getKind() != ExprKind::DeclRef)
-        return Action::Continue(E);
-      auto D = dyn_cast<DeclRefExpr>(E)->getDecl();
-      if (D->getKind() != DeclKind::Var && D->getKind() != DeclKind::Param)
-        return Action::Continue(E);
-      VarName = dyn_cast<VarDecl>(D)->getName().str().str();
-      return Action::Stop();
-    }
-  };
-
-  class ConditionalPatternFinder : public ASTWalker {
-  public:
-    ConditionalPatternFinder(SourceManager &SM) : SM(SM) {}
-
-    SmallString<64> ConditionalPattern = SmallString<64>();
-
-    MacroWalking getMacroWalkingBehavior() const override {
-      return MacroWalking::Arguments;
-    }
-
-    PostWalkResult<Expr *> walkToExprPost(Expr *E) override {
-      auto *BE = dyn_cast<BinaryExpr>(E);
-      if (!BE)
-        return Action::Continue(E);
-      if (isFunctionNameAllowed(BE))
-        appendPattern(BE->getLHS(), BE->getRHS());
-      return Action::Continue(E);
-    }
-
-    PreWalkResult<Pattern *> walkToPatternPre(Pattern *P) override {
-      ConditionalPattern.append(Lexer::getCharSourceRangeFromSourceRange(SM, P->getSourceRange()).str());
-      if (P->getKind() == PatternKind::OptionalSome)
-        ConditionalPattern.append("?");
-      return Action::Stop();
-    }
-
-  private:
-
-    SourceManager &SM;
-
-    bool isFunctionNameAllowed(BinaryExpr *E) {
-      Expr *Fn = E->getFn();
-      if (auto DotSyntaxCall = dyn_cast_or_null<DotSyntaxCallExpr>(Fn)) {
-        Fn = DotSyntaxCall->getFn();
-      }
-      DeclRefExpr *DeclRef = dyn_cast_or_null<DeclRefExpr>(Fn);
-      if (!DeclRef) {
-        return false;
-      }
-      auto FunctionDeclaration = dyn_cast_or_null<FuncDecl>(DeclRef->getDecl());
-      if (!FunctionDeclaration) {
-        return false;
-      }
-      auto &ASTCtx = FunctionDeclaration->getASTContext();
-      const auto FunctionName = FunctionDeclaration->getBaseIdentifier();
-      return FunctionName == ASTCtx.Id_MatchOperator ||
-             FunctionName == ASTCtx.Id_EqualsOperator ||
-             FunctionName == ASTCtx.Id_derived_enum_equals ||
-             FunctionName == ASTCtx.Id_derived_struct_equals;
-    }
-
-    void appendPattern(Expr *LHS, Expr *RHS) {
-      auto *PatternArgument = RHS;
-      if (PatternArgument->getKind() == ExprKind::DeclRef)
-        PatternArgument = LHS;
-      if (ConditionalPattern.size() > 0)
-        ConditionalPattern.append(", ");
-      ConditionalPattern.append(Lexer::getCharSourceRangeFromSourceRange(SM, PatternArgument->getSourceRange()).str());
-    }
-  };
-
-  class ConverterToSwitch {
-  public:
-    ConverterToSwitch(const ResolvedRangeInfo &Info,
-                      SourceManager &SM) : Info(Info), SM(SM) { }
-
-    void performConvert(SmallString<64> &Out) {
-      If = findIf();
-      OptionalLabel = If->getLabelInfo().Name.str().str();
-      ControlExpression = findControlExpression();
-      findPatternsAndBodies(PatternsAndBodies);
-      DefaultStatements = findDefaultStatements();
-      makeSwitchStatement(Out);
-    }
-
-  private:
-    const ResolvedRangeInfo &Info;
-    SourceManager &SM;
-
-    IfStmt *If;
-    IfStmt *PreviousIf;
-
-    std::string OptionalLabel;
-    std::string ControlExpression;
-    SmallVector<std::pair<std::string, std::string>, 16> PatternsAndBodies;
-    std::string DefaultStatements;
-
-    IfStmt *findIf() {
-      auto S = Info.ContainedNodes[0].dyn_cast<Stmt*>();
-      return dyn_cast<IfStmt>(S);
-    }
-
-    std::string findControlExpression() {
-      auto ConditionElement = If->getCond().front();
-      auto Finder = VarNameFinder();
-      ConditionElement.walk(Finder);
-      return Finder.VarName;
-    }
-
-    void findPatternsAndBodies(SmallVectorImpl<std::pair<std::string, std::string>> &Out) {
-      do {
-        auto pattern = findPattern();
-        auto body = findBodyStatements();
-        Out.push_back(std::make_pair(pattern, body));
-        PreviousIf = If;
-      } while ((If = dyn_cast_or_null<IfStmt>(If->getElseStmt())));
-    }
-
-    std::string findPattern() {
-      auto ConditionElement = If->getCond().front();
-      auto Finder = ConditionalPatternFinder(SM);
-      ConditionElement.walk(Finder);
-      return Finder.ConditionalPattern.str().str();
-    }
-
-    std::string findBodyStatements() {
-      return findBodyWithoutBraces(If->getThenStmt());
-    }
-
-    std::string findDefaultStatements() {
-      auto ElseBody = dyn_cast_or_null<BraceStmt>(PreviousIf->getElseStmt());
-      if (!ElseBody)
-        return getTokenText(tok::kw_break).str();
-      return findBodyWithoutBraces(ElseBody);
-    }
-
-    std::string findBodyWithoutBraces(Stmt *body) {
-      auto BS = dyn_cast<BraceStmt>(body);
-      if (!BS)
-        return Lexer::getCharSourceRangeFromSourceRange(SM, body->getSourceRange()).str().str();
-      if (BS->getElements().empty())
-        return getTokenText(tok::kw_break).str();
-      SourceRange BodyRange = BS->getElements().front().getSourceRange();
-      BodyRange.widen(BS->getElements().back().getSourceRange());
-      return Lexer::getCharSourceRangeFromSourceRange(SM, BodyRange).str().str();
-    }
-
-    void makeSwitchStatement(SmallString<64> &Out) {
-      StringRef Space = " ";
-      StringRef NewLine = "\n";
-      llvm::raw_svector_ostream OS(Out);
-      if (OptionalLabel.size() > 0)
-        OS << OptionalLabel << ":" << Space;
-      OS << tok::kw_switch << Space << ControlExpression << Space << tok::l_brace << NewLine;
-      for (auto &pair : PatternsAndBodies) {
-        OS << tok::kw_case << Space << pair.first << tok::colon << NewLine;
-        OS << pair.second << NewLine;
-      }
-      OS << tok::kw_default << tok::colon << NewLine;
-      OS << DefaultStatements << NewLine;
-      OS << tok::r_brace;
-    }
-
-  };
-
-  SmallString<64> result;
-  ConverterToSwitch(RangeInfo, SM).performConvert(result);
-  EditConsumer.accept(SM, RangeInfo.ContentRange, result.str());
   return false;
 }
